Use long for episode number

+++ TMDbLib/Objects/Search/SearchTvEpisode.cs
@@ -27,7 +27,7 @@ public class SearchTvEpisode : SearchBase
     /// Gets or sets the episode number.
     /// </summary>
     [JsonProperty("episode_number")]
-    public int EpisodeNumber { get; set; }
+    public long EpisodeNumber { get; set; }
 
     /// <summary>
     /// Gets or sets the name of the episode.
