+++ TMDbLib/Objects/TvShows/TvEpisodeInfo.cs
@@ -23,5 +23,5 @@ public class TvEpisodeInfo
     /// Gets or sets the episode number.
     /// </summary>
     [JsonProperty("episode_number")]
-    public int EpisodeNumber { get; set; }
+    public long EpisodeNumber { get; set; }
 }
